requirements-core.txt
boto3==1.14.20
csscompressor==0.9.5
django-compressor==2.4
django-storages==1.9.1
Django==3.0.8
mock==4.0.2
selenium==3.141.0
psycopg2-binary==2.8.5
pytz==2020.1

# For coverage
autopep8==1.5.3
coverage==5.2
django-nose==1.4.6
pylint==2.5.3
pylint-django==2.2.0

# For django-rest
djangorestframework==3.11.0
markdown==3.2.2

# For movie creation
celery==4.4.6
django-cleanup==5.0.0
moviepy==1.0.3
